Eliminate client code from this repository and refactor.
The folder application/client has been removed. The client code
was packaged and placed here for reuse:

https://github.com/agentsofthesystem/operator

The AgentSmith software codebase was refactored to use the new
package.

Please note, for now, the requirements.txt points to the main
branch of the client instead of a fixed release or tag. In the
future, discrete versions should be used.