Correct performance issue with GM GUI and associated bugs.
There were bugs preventing the window from refreshing properly. Those
have been corrected.

While the window was closed, it was also refreshing in the background
and making other windows/widget slow.  So, the refreshing feature
only happens now when the window is open.  The timer behind it will
start/stop with respect to the window being opened or not.